move join.py to back whence it came
